SW-4330-continued Allow reassign-request modal to show in full screen
- Created a custom hook that informs when a map is in fullscreen and which container to use for modal portals
- Integrated this hook with map tooltip so reassign request modal shows in full screen mode as well (uses the full screen container as the parent for the modal)
- popovers (such as layers and outdoors-vs-satellite) are still not functional, added a change to not render these in fullscreen mode for now (since they won't work)
- will follow up later with a web-component change
